requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_huggingface==0.3.1
langchain_openai==0.3.30
pydantic==2.11.7
python-dotenv==1.1.1
sentence-transformers==5.0.0
uvicorn==0.35.0